backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
ortools==9.8.3296
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiofiles==23.2.1
